polymarket/requirements.txt
py-clob-client
aiohttp
python-dotenv
termcolor
